check for source and target directory is the same

+++ egs/wsj/s5/utils/format_lm.sh
@@ -36,13 +36,15 @@ mkdir -p $out_dir
 
 echo "Converting '$lm' to FST"
 
-if [ -e $out_dir/phones ]; then
-  rm -r $out_dir/phones
-fi
+if [ -e $out_dir ] && [ ! $lang_dir -ef $out_dir ] ; then
+  if [ -e $out_dir/phones ] ; then
+    rm -r $out_dir/phones
+  fi
 
-for f in phones.txt words.txt topo L.fst L_disambig.fst phones oov.int oov.txt; do
-  cp -r $lang_dir/$f $out_dir
-done
+  for f in phones.txt words.txt topo L.fst L_disambig.fst phones oov.int oov.txt; do
+     cp -r $lang_dir/$f $out_dir
+  done
+fi
 
 lm_base=$(basename $lm '.gz')
 gunzip -c $lm \
